requirements.txt
streamlit
pyautogen
python-dotenv
pandas
qrcode[pil]
geopy
seaborn
matplotlib
streamlit-folium
folium